backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
Django==3.1.6
django-cors-headers==3.7.0
django-material-admin==1.7.16
django-rest-framework==0.1.0
djangorestframework==3.12.2
filelock==3.0.12
greenlet==1.0.0
jedi==0.18.0
msgpack==1.0.2
neovim==0.3.1
parso==0.8.1
Pillow==8.1.0
psycopg2==2.8.6
pynvim==0.4.3
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
